// frontend/src/types/api.ts
export interface ProductCreate {
  operator_id?: number | null;
  name: string;
  sku?: string | null;
  category?: string | null;
  unit_cost?: number | string | null;
  retail_price?: number | string | null;
  par_level?: number | null;
  image_url?: string | null;
  is_active?: boolean | null;
  [key: string]: any;
}
export interface ProductResponse {
  operator_id?: number | null;
  name: string;
  sku?: string | null;
  category?: string | null;
  unit_cost?: string | null;
  retail_price?: string | null;
  par_level?: number | null;
  image_url?: string | null;
  is_active?: boolean | null;
  id: number;
  created_at?: string | null;
  updated_at?: string | null;
  [key: string]: any;
}
export interface ProductUpdate {
  operator_id?: number | null;
  name?: string | null;
  sku?: string | null;
  category?: string | null;
  unit_cost?: number | string | null;
  retail_price?: number | string | null;
  par_level?: number | null;
  image_url?: string | null;
  is_active?: boolean | null;
  [key: string]: any;
}